comments re: prettier, naming, versions
