+++ src/components/EntityField.tsx
@@ -1,4 +1,4 @@
-import {createContext, useContext, useState} from "react";
+import { createContext, useContext, useState } from "react";
 
 type EntityFieldContextProps = {
   tooltipsVisible: boolean;
