+++ src/components/index.ts
@@ -1,2 +1,2 @@
-export { Editor, TARGET_ORIGINS } from "./editor.tsx";
-export {EntityFieldProvider, useEntityField} from "./EntityField.tsx"
+export { Editor, TARGET_ORIGINS } from "./Editor.tsx";
+export { EntityFieldProvider, useEntityField } from "./EntityField.tsx";
