+++ src/internal/puck/ui/Progress.tsx
@@ -10,7 +10,7 @@ const Progress = React.forwardRef<
     ref={ref}
     className={cn(
       "relative h-4 w-full overflow-hidden rounded-full bg-secondary",
-      className,
+      className
     )}
     {...props}
   >
